fix integer type

+++ search_algorithms/1-binary.c
@@ -8,10 +8,10 @@
  */
 int binary_search(int *array, size_t size, int value)
 {
-	size_t low = 0;
-	size_t high = size - 1;
-	size_t mid = low + (high - low) / 2;
-	size_t i;
+	int low = 0;
+	int high = size - 1;
+	int mid = low + (high - low) / 2;
+	int i;
 
 	if (array == NULL)
 		return (-1);
